NJ tree building fixed selecting nodes to merge
Added a proper arg_min function to select randomly one of the node
pairs with the shortest distance. Now NJ tree doesn't glitch out if
sequences are the same and there are a lot of 0 cells in the Q matrix.
Added more NJ tests: a random web example and an example based on a
protein sequence file with just one change in a single sequence.