Parsimony alignment fix (#2)
* Minor fixes to make linter happy

Minor fixes to make clippy happy, removed unnecessary vec!, into_iter()
and the like.

* Fixed sequence order to match leaves

Fixed sequence order in the vector to match leaf indices. The code relied
on that property to begin with, but it was neither being tested nor
was it being enforced in any way which lead to very malformed alignments
in IndelMaP.
This contains the fix and a test to check that the order matches together
with the data for the test.