Added struct for defining rounding
Added a struct that contains 2 values -- whether to round some of the numbers
and if yes, to how many decimal digits. Not necessary in general, but used in
testing against the values produced by the python scripts, so now rounding is
optional for parsimony scores derived from the substitution models and for the
branch percentiles.