Protein substitution matrix transpose fix (#3)
* Added struct for defining rounding

Added a struct that contains 2 values -- whether to round some of the numbers
and if yes, to how many decimal digits. Not necessary in general, but used in
testing against the values produced by the python scripts, so now rounding is
optional for parsimony scores derived from the substitution models and for the
branch percentiles.

* Added NodeIdx display and node id printing

Implemented Display for NodeIdx so that it always prints what kind of node
it is.
Added a function that helps with logging -- generates string "with ID xxx"
if there is an ID attached to the current node, or gives back an empty string.

* Fixed protein matrices from by row to by cols

The provided protein substitution matrices were actually given by rows
whereas the Matrix struct reads them by columns.
Transposed the matrices to match proper order of cols/rows.

* Added tests for output to appease codecov

Added tests for the helper functions to appease codecov so it lets me merge.