Changed FreqVector and SubstMatrix to dynamic size
Changed the FreqVector and the SubstMatrix type to dynamic sizing to make them usable
when the parametrisation on N (number of chars) becomes different for a model with gaps.
It also makes sure that all the data types we use are the same, no real point in using
statically typed matrices when most will still be dynamically typed (e.g. the partial
likelihood matrices).