Indelmap compatibility fix (#8)
* Update PhyloInfo creator functions

Updated PhyloInfo creator functions to allow proper initialisation checks when
creating PhyloInfo from a vector of sequences and a newick string or an existing tree.
Using the new functions in all tests now.
Previously the checks were only performed when initialising from files.

* Likelihood setup functions made public

Made the likelihood setup functions public (as they should be, and avoids
dead code warnings)

* Moved creating parsimony scorings to diff trait

Moved scoring matrix creation for substitution models to a trait different from
EvolutionaryModels -- since it does not make sense to create scoring matrices for
PIP models. Now they are in a separate trait called ParsimonyModel.

* Test for phyloinfor from sequences and newick

* Tests for phyloInfo phyloinfo_from_sequences_tree

* Removed extra whitespace to please fmt

* tree::new now returns result to handle one leaf

Tree::new function now returns Result<Tree> to capture the case when
there are zero sequences provided. It also manages when a single
sequence is given by setting it as the only leaf and the root.
Multiple sequences are handled as before.

* Added tests for likelihood on single node and fail

Added two new tests:
    A test for when we try to create the likelihood function without
an alignment -- should return an error.
    A test for when we only have a single node in the tree, should
compute the likelihood anyway.

* Fixed bug in param counts for hky and tn93

Fixed bad error messages for cases when there is a wrong number of
parameters given for hky and tn93 models -- the check had 10 parameters
for both models, copy paste error from gtr.

* Added tests to call EvoModel methods

Added tests to use EvolutionaryModel trait methods for DNASubstitutionModel,
get_rate and get_stationary_distribution.

* Added test to call ParsimonyModel methods for DNA

Added tests to run ParsimonyModel::generate_scorings and ParsimonyModel::get_scoring_matrix
for DNA models.

* Chang sequence vec to slice in tree creation

Removed unnecessary vectors from tree creation, the sequences can
be provided as a slice.

* Added test for Tree::new with no sequences

The tree should not be created if Tree::new is given an empty slice
of sequences.

* Test get_idx_by_id and fix panic when not found

Added a test for the det_idx_by_id function in the tree.
Changed function signature to return Result with the idx if the id
exists and an Error otherwise.

* PhyloInfo sequences should be pub

PhyloInfo sequences should be pub for the parsimony aligner. changed
visibility.