Subst models always normalised to avg rate of 1
Substitution models are now all automatically normalised to an average rate of 1.
There used to be an option to noemalise/not, but in order for the PIP deletion rates to
be sensible w.r.t. the substitution rates it makes sense to fix the rate prior to setting
up PIP.
Therefore removed the normalisation function from EvolutionaryModel trait, and also
added a reset method to EvolutionaryModelInfo to allow reusing temp storage.