Subst model parameter optimisation fix (#11)
* Added option to give only kappa to k80

Added the option to provide a single parameter to k80 (kappa), which is equivalent to setting
alpha to kappa and beta to 1.

* Overhaul for DNA model creation

Cleaned up DNA model creation and made sure the unnormalised models
use all the right parameters and don't normalise under the hood because
this messes with the creation of PIP.

* Update likelihood calculation to use external tmp

Updated the likelihood computation to not store the temporary values inside itself
as that makes it necessary for object ot be mutable when computing the likelihood,
which is not compatible with the argmin library. Now the storage has to be provided
extrenally.

* Subst models always normalised to avg rate of 1

Substitution models are now all automatically normalised to an average rate of 1.
There used to be an option to noemalise/not, but in order for the PIP deletion rates to
be sensible w.r.t. the substitution rates it makes sense to fix the rate prior to setting
up PIP.
Therefore removed the normalisation function from EvolutionaryModel trait, and also
added a reset method to EvolutionaryModelInfo to allow reusing temp storage.

* K80 model parameter optimiser using brent

Added a substitution model parameter optimiser using one-dim brent on alpha and beta
sequentially to get estimates.

* Fixed filename typo in K80 param optim tests

* DNA model creation partial refactor

DNA model creation simplified to not have to use str model names everytime.
Moved K80-related functions/structures to separate module, added pub use statement to
dna_models module

* Refactored jc69 to separate module

* Refactored hky to separate module

* Refactored tn93 to separate module

* Refactored gtr to separate module.

Final refactor of DNA substitution models into separate modules.

* PhyloInfo makes sequences uppercase on creation

PhyloInfo now makes all sequences uppercase when created using either
of the three available creation methods, from files, from a slice of Records and
a tree, or from a slice of records and a newick string.

* Add docs and doctests to pub methods in PhyloInfo

Added doctrings to everything and doctests to public methods in PhyloInfo.

* Removed redundant PhyloInfo creation function

Removed the redundant PhyloInfo creation function phyloinfo_from_sequences_newick,
phyloinfo_from_sequences_tree can be used instead.

* Add method to get empirical freqs from PhyloInfo

Added a method to get the empirical character frequencies from PhyloInfo.

* Doctests for io

Added doctests for sequence and tree io.

* Added a check for existing fasta file on output

Added a check to make sure that if a file exists it wont be overwritten when writing
sequences. Later should add an option to specify if overwriting is allowed, but by default
should not be.

* Removed redundant stat frequency check

Removed the check for stationary frequencies summing up to 1 because it is already
done in the make_pi method.

* Test fixes ot empirical frequencies in phyloInfo

Test fixes for PhyloInfo get_empirical frequencies

* Added method for emp freqs for evo models

Added a method for getting the empirical frequencies ffor data for evolutionary models.
Implemented for DNA substitutions, stubs for all else for now.

* Tree parser made public and use cleanup

Made the tree_parser public for parsing newicks, cleaned up use statements

* Tests for DNA emp frequencies with ambig chars

Tests for getting correct DNA empirical frequencies for all the different ambiguous characters

* Method name fix in a doctest

* Added HashMap representing DNA ambig chars

Added a HashMap that is used to represent ambiguous DNA characters that can be reused in
other places so that the ambiguous characters are always fixed and exist only in one place

* DNA models parametrised with same struct

Replaced all the random parametrisations of different DNA models with a single
struct to use for optimisation later.

* PhyloInfo returns char counts

PhyloInfo now returns different character counts rather than frequencies so that in stationary
frequency estimation no frequency accidentally becomes 0.0 -- now they are adjusted by a single
char "added" to the sequences.

* Initial optim of GTR

Initial attempt at optimising GTR parameters, very slow and untested for now.

* Using parameter struct for models instead of vec

Replaced the parameter vector that each model stores with the appropriate
parameter struct.

* GTR param optimisation v1

GTR params optimised inefficiently with gold section search and with rag fixed to 1.0.

* Fixed bug for incorrect likelihood

The temp values were not being cleared sometimes when computing likelihoods.
Now the temp storage is initialised in the likelihood computation through the
trait functions and the internal implementation can use/reuse the temp storage
rather than forcing this to be always the case.

* Refactoring for GTR optimisation

Refactoring GTR optimisation to reduce unnecessary model copying all over the place.
Passing parameter values instead.

* DNA model optimisation with uniform framework

DNA model parameter optimisation now done in a uniform framework for all models.
Still very slow and needs profiling.

* Fixed creating an expensive hashmap

Reduced GTR parameter optimisation from 1:30 to 24s by making the ambiguous char
mapping static.
Next will remove hashmap entirely, it's not needed here and it's insanely slow.

* Removed HashMap for ambig chars, array instead

Removed HashMap from any computations involving ambiguous chars, now using statically
computed arrays.

* Proper leaf encoding handling with static arrays

Two major changes:
- include a GapHandling enum option for handling gaps as ambig chars for computing
likelihoods without PIP
- create static arrays to get character encodings for tree leaves to save time, these
indices will not change and the encodings should not be recomputed on the same MSA.

* changed gtr visibility to pub for library

* Release level optimisation set to max

* Fixed unnecessary reference to a Vec

Replaced unnecessary &Vec with &[], older local version of clippy did not complain about it
but an updated one on girhub did.

* Replace println with logging, fix typo

* Change more &Vec to &[] as per clippys complaints

* Change &Vec to &[] in tests

* Removed unnecessary helper function

* PhyloInfo: added sequence_type to attributes

Added sequence_type as an attribute to PhyloInfo as the sequence type is determined on read and
will remain the same throughout the analysis;
Removed alphabet parameter from get_counts method.

* PhyloInfo consumes Vec of Records rather than copy

The phyloinfo_from_sequences_tree function now consumes the vec of sequences
rather than copying it since the copy was useless;
Refactor to use this function from phyloinfo_from_files to avoid dublicate code.

* Added todo for protein data

* PhyloInfo: add methods to read sequences w/o tree

Added two methods:
    - takes vec of sequences and makes an NJ tree;
    - reads sequences from file then runs prev method.

* PhyloInfo generation function refactor

PhyloInfo generation functions refactored to be within the PhyloInfo class,
removed duplicate naming.

* trying to make rustfmt work again

* trying to make rustfmt work again

* Format Rust code using rustfmt

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>