Refactor alignment to use builder, class for seqs
Refactored alignment creation into a builder struct, the internal representation
can be built from either a given MSA or from a map of internal nodes.
Now the alignment class also stores the ungapped sequences and, for a current alignment,
the maps of leaf sequences to gaps/characters.
Sequences have a separate struct for storage that has several methods to simplify access.