Add scientific floats to newick tree parser (#38)
* Allowing underscore and asterisk as gap

* Filtering out empty columns from MSA

* Add test for pip on an msa w/ empty col

Added a test for PIP on an MSA woth an empty column: this used to cause the log
likelihood to become -Inf, now such columns are filtered out by the alignment builder
and no longer cause issues.

* Added missing test files

* Fix wrong alphabet after removing gap cols

Fixed screwing up the alphabet when gap columns are removed

* Add Display to alphabet

* Test for alphabet name

* Remove warn! to appease coverage gods

* Fix test file to report correct error

* Add error check to test with invalid fasta

* Add scientific floats to newick grammar

Added scientific float representation to the newick parsing expression grammar.
Now numbers like 3.4e-10 or .0003 are valid and are parsed correctly.

* Add test for newick with scientific floats