Alignment refactor (#41)
* Replace PartialEq to compare regardless of order

Replaced derived PartialEq with an implementation that ignores the order in which the
sequences are stored internally.

* Remove compile_from_map and related tests

Remove compile_from_map from AlignmentBuilder as it was an internal function
which too strongly depends on creating the maps correctly. Removed as redundant and
confusing.

* Refactor compile alignment to use Sequences

Refactored the compile alignment methods to return a Sequences object
rather than a vector of records for no reason.

* Add Display to Sequences object

* impl Display for Alignment and Sequences

* Tests for Display in Sequences and Alignment

* Fix rust fmt

Setting rust fmt check to an older version as the new one seems to be broken

* Fix rust fmt again