Trying to make getting costs per branch efficient
Using a Vec<(OrderedFloat<f64>,BranchModelCost)> to store costs per branch length and binary search
for the closest one when requenced. Seems to be the fastest way to do this given that we have a predefined
set of times that we need to have individual scorings for, but branch lengths may change.