Set probabilities at root to f64::min if zero
Hack to avoid running into underflow when branch lengths grow to infinity and
site likelihoods become 0. If at the root the individual probability of a single site
is 0.0, set it to f64::MIN_POSITIVE instead to limit the likelihood to a numerical
value. Needs to be tested for stability/numerical correctness.