Remove unused `tree_builder` field
Removed unused `tree_builder` field from `PhyloInfoBuilder`, since there is no
way to actually set it at the moment, meaning it always defaults to None. However, the
`build_nj_tree` method would still try to use it but did not check if the distance measure is
valid for the sequences. This means that if the option to set a tree builder were added,
the `build_nj_tree` could behave unexpectedly. It is safer to entirely remove this field for now.