Fix inconsistent saving of costs in ModelOptimiser
Fix inconsistent saving of costs in ModelOptimiser -- after adding a costs: Vec<f64> field
to ModelOptimisationResult, the first cost pushed into the vector was not the cost at the start
but rather the cost after the potential frequency optimisation which was not consistent behaviour
with other optimisers. Also added the check for the first cost value to tests.